Server Auth implementation (to be tested) + small fix for BlockStoreAuthEnc
Server Auth remains to be tested. Testing is a gigantic pain. Apparently each block has access to a virtual "superblock" of size 64 bytes that is actually a regular block. I saw this while trying to understand multiplexor. Similarly, our AuthEnc Store is used to set up the Multiplexor so any operation on it is regimented by our algorithm. It has yet to be seen if there are collisions in the use of memory in the superblock.